Update AsyncWorker.hh
Want to see if when I join m_worker I am getting std::system_error because of resource_deadlock, which might be caused because the system thinks I am stuck in an infinite loop (even though I have a break). So now waiting till I break out to call to m_worker.join(). Also now done really is the end so don't have to expose this to the child destructor.